MAJOR RESTRUCTURING OF views FOLDER
