+++ src/views/AdHocRecipe/Recipe.vue
@@ -44,7 +44,7 @@
 
 <script>
 import axios from 'axios'
-import AnimatedPlaceholder from '../components/AnimatedPlaceholder.vue'
+import AnimatedPlaceholder from '../../components/AnimatedPlaceholder.vue'
 
 export default {
     data() {
