--- src/views/Preferences.vue
@@ -1,18 +1,0 @@
-<template>
-    <main class="row d-flex justify-content-center">
-        <div class="col-4">
-            <div className="d-flex align-items-center justify-content-center" style="height:100vh">
-                <searchBar/>
-                <searchBar/>  
-            </div>
-        </div>
-    </main>
-</template>
-
-<script>
-    
-</script>
-
-<style scoped>
-
-</style>
